Time: 4 ms (82.77%) | Memory: 42.1 MB (68.45%) - LeetSync

+++ 633-sum-of-square-numbers/sum-of-square-numbers.java
@@ -6,7 +6,7 @@ class Solution {
         while (a <= b) {
             long sumOfsquares = (long) (a * a) + (b * b);
 
-            if (sumOfsquares == (long) c) {
+            if (sumOfsquares == c) {
                 return true;
             } else if (sumOfsquares < c) {
                 a++;
